requirements.txt
psycopg2-binary==2.8.6
Flask==1.1.2
requests==2.25.0
beautifulsoup4==4.9.3
configparser==5.0.1
